backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
pydantic
sqlalchemy
asyncpg
aiosqlite
greenlet
